requirements.txt
flask
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
prophet
requests
pillow
google-generativeai